Refactor RosterMatcher into salt.utils
